tests/submission/PhilipTaylor/canvas/2d.state.saverestore.stackdepth.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.state.saverestore.stackdepth</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.state.html">state</a>.<a href="index.2d.state.saverestore.html">saverestore</a>.stackdepth</h1>
<p class="desc">save()/restore() stack depth is not unreasonably limited</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.state.save">2d.state.save</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.state.restore">2d.state.restore</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() stack depth is not unreasonably limited");
_addTest(function(canvas, ctx) {

var limit = 512;
for (var i = 1; i < limit; ++i)
{
    ctx.save();
    ctx.lineWidth = i;
}
for (var i = limit-1; i > 0; --i)
{
    _assertEqual(ctx.lineWidth, i, "ctx.lineWidth", "i");
    ctx.restore();
}


});
</script>

<div id="log"></div>
</body></html>
